// types/Ticketing.type.ts
export type TicketListItemInterface = {
    ticket_id: string;
    title: string;
    description: string;
    priority: string;
    status: string,
    created_at: string;
    updated_at: string;
    created_by: string;
    assigned_to: string;
    comments: string;
    collaborators: string;
    screenshot: string;
    department: string;
};